moved files into schema and switched extensions
