--- Pop-Grupa-A-Frontend/src/app/common/alert/StyledAlert.js
@@ -1,8 +1,0 @@
-import styled from 'styled-components';
-
-export const StyledAlert = styled.div`
-  width: 40%;
-  height: 80px;
-  margin: 20px auto;
-  
-`;
