--- Pop-Grupa-A-Frontend/src/app/login/component/LoginForm.css
@@ -1,8 +1,0 @@
-.login-form {
-    width: 30%;
-    margin: 0 auto 0 auto;
-}
-
-.login-form div {
-    margin-bottom: 10px;
-}
