+++ Pop-Grupa-A-Frontend/src/components/common/NotFoundPage.jsx
@@ -2,7 +2,9 @@ import React from 'react';
 
 const NotFoundPage = () => (
     <React.Fragment>
-        Not Found the page you are looking for.
+        <h2>
+            Not Found the page you are looking for.
+        </h2>
     </React.Fragment>
 );
