+++ Pop-Grupa-A-Frontend/src/components/common/StyledCard.jsx
