+++ Pop-Grupa-A-Frontend/src/redux/alert/alertReducer.js
@@ -2,15 +2,16 @@ import types from './alertTypes';
 
 const INITIAL_STATE = {
     show: false,
-    message: ''
+    message: '',
+    title: ''
 };
 
 const alertReducer = (state = INITIAL_STATE, action) => {
     switch(action.type) {
         case types.SHOW_ALERT:
-            return {...state, show: true, message: action.payload};
+            return {...state, show: true, message: action.payload.message, title: action.payload.title};
         case types.HIDE_ALERT:
-            return {...state, show: false, message: ''};
+            return {...state, show: false, message: '', title: ''};
         default:
             return state;
     }
