+++ Pop-Grupa-A-Frontend/src/redux/routing/ProtectedRoutes.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import {Route} from "react-router-dom";
 import {Redirect} from 'react-router-dom';
-import {UserIdPath, RolePath } from "../../constants";
+import {UserIdPath, RolePath } from "../../commons/Constants";
 import * as Cookies from 'js-cookie';
 
 const ProtectedRoute = ({component: Component, roles, ...other}) => {
